requirements.txt
pandas==2.2.2
openpyxl==3.1.5
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
google-generativeai==0.7.2
cloudinary==1.40.0
Pillow==10.4.0
tenacity==8.3.0
